feat(main): buttonコンポーネントのリファクタリングを行う
