+++ app/components/toppage/elements/disclaimer-button.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import { useState } from "react";
 import Modal from "react-modal";
 import { MdContentCopy } from "react-icons/md";
-import { IoMdCheckmarkCircleOutline } from "react-icons/io";
+import { useModal } from "@/app/hooks/toppage/useModal";
+import { DisClaimerInModalContent } from "./inmodal/disclaimer-in-modal-content";
 
 const modalStyle = {
   overlay: {
@@ -25,9 +25,7 @@ const modalStyle = {
 };
 
 export function DisclaimerButton() {
-  const [modalIsOpen, setIsOpen] = useState<boolean>(false);
-  const modalOpen = () => setIsOpen(true);
-  const modalClose = () => setIsOpen(false);
+  const { isModalOpen, modalOpen, modalClose } = useModal();
 
   return (
     <>
@@ -38,38 +36,9 @@ export function DisclaimerButton() {
         <MdContentCopy className="mr-2" />
         Disclaimer
       </div>
-      <Modal isOpen={modalIsOpen} style={modalStyle}>
-        <InModalContent modalCloseFunc={modalClose} />
+      <Modal isOpen={isModalOpen} style={modalStyle}>
+        <DisClaimerInModalContent modalCloseFunc={modalClose} />
       </Modal>
-    </>
-  );
-}
-
-// TODO:下のコンポーネントをファイルで分けるかどうか迷っている
-// 分けるメリット:1ファイル1コンポーネントが守られていて、心理的に見やすい
-// 分けるデメリット:elementsの中にlayoutから直接参照されていないファイルが出来てしまう、ディレクトリ構造としてはパッと見で分かりずらい
-
-interface InModalContent {
-  modalCloseFunc: () => void;
-}
-
-export function InModalContent({ modalCloseFunc }: InModalContent) {
-  return (
-    <>
-      <p className="font-bold text-lg">免責事項</p>
-      <p className="text-gray-600 mt-2">
-        このウェブアプリを利用することによって生じるあらゆる損害や損失に対して、一切の責任を負いません
-      </p>
-      <hr className="mt-2 h-2 mb-1" />
-      <button
-        className="w-full rounded-lg bg-green-700 hover:bg-green-800 py-2 transition-colors duration-300"
-        onClick={modalCloseFunc}
-      >
-        <div className="flex items-center justify-center text-white">
-          <IoMdCheckmarkCircleOutline size={20} className="mr-2" />
-          同意する
-        </div>
-      </button>
     </>
   );
 }
