+++ app/components/toppage/elements/inmodal/disclaimer/modal-content.tsx
@@ -1,0 +1,7 @@
+export function ModalContent() {
+  return (
+    <p className="text-gray-600 mt-2">
+      このウェブアプリを利用することによって生じるあらゆる損害や損失に対して、一切の責任を負いません
+    </p>
+  );
+}
