+++ app/components/toppage/elements/inmodal/disclaimer/modal-header.tsx
@@ -1,0 +1,3 @@
+export function ModalHeader() {
+  return <p className="font-bold text-lg">免責事項</p>;
+}
