+++ app/components/toppage/elements/inmodal/main/modal-content.tsx
@@ -1,0 +1,7 @@
+export function ModalContent() {
+  return (
+    <div className="text-gray-500 mt-2">
+      <p>EnvHubの利用にはGitHubアカウントでのログインが必要です。</p>
+    </div>
+  );
+}
